requirements-docker.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-generativeai==0.8.5
python-dotenv==1.0.1
python-multipart==0.0.6
pydantic==2.11.7